Crate expectest

Crate provides matchers and matcher functions for unit testing.

Modules

core
A module contains core types of this library.

matchers
A module contains matchers.

prelude
A module contains reexport of all useful functions.

Macros

expect!
A macro intended to use instead of expect function.