Module expectest::core

A module contains core types of this library.

Structs

ActualValue
Represent an actual value and optional location of a test case in a source code.

Failure
Represents data for Failure variant of a TestResult.

SourceLocation
Represents a location in a source code.

Enums

Join
An enum used to join two parts of a failure message.

TestResult
Represents result of unit testing.

Traits

Matcher
A Matcher trait.

Functions

expect
Creates a new instance of ActualValue using value.