Enum expectest::core::TestResult

pub enum TestResult {
    Success,
    Failure(Failure),
}

Represents result of unit testing.

Variants

Success

Denotes success.

Failure

Denotes failure.

Methods

impl TestResult

fn new_success() -> Self

Creates a new instance of TestResult with a Success variant.

fn new_failure(message: String, location: Option<SourceLocation>) -> Self

Creates a new instance of TestResult with a Failure variant using message and location.

fn assert_eq_message(self, message: &str)

Asserts that message is equal to the failure message.

Purpose of this method is the unit testing of failure messages. Convenient for development.

Example

expect(0 == 1).to(be_true()).assert_eq_message("expected to be true");

Panics

This method panics if the failure message is not equal to the message, or self is not the TestResult::Failure variant.